backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
langchain==0.2.1
langchain-community==0.2.1
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
chromadb==0.5.0
google-genai>=1.0.0
groq>=0.9.0
pypdf==4.2.0
